+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [4.3.8.X] - XXXX-XX-XX
 ### Added
 - View menu for navigating main views alongside key shortcuts (and circular navigation)
-
+- Show a menu for the server console in the menu bar on the top
 
 ## [4.3.7.3] - 2025-03-24
 ### Fixed